Redo Reader.
This started as a move to streams. Then, when implementing the functions 
to read from a string, after attempting a lookup table, a new case 
statement was written. Then types were overhauled.

This ends up completely redoing the Reader. Support is preserved for 
VarInt and strings (user-defined length delimited types are currenty 
untested). Zigzags ints are broken with them being the next target.

Once the Reader is re-implemented, the API of test_serialization will be 
updated. If it passed, meaning the Reader is valid, the Writer will be 
re-implemented.