Merge bit distinctions for sint/pint. Remove subtypes where possible.
This is really a cleanup commit.

Since we switch off the type, and error when the wrong pragma (sint32 
for int64) was attached, it's easy enough to make it so you can't attach 
the wrong pragma. Now it's just pint/sint.

Subtypes are only needed for VarInt. The only other ambiguous type is 
caused by an int/uint being able in a fixed format, yet due to the 
different wire type, we don't need to determine a subtype for it.