Remove any guessing/defaulting to what type to use.
We only defaulted when handed a signed integer outside of an object, 
with no context on how to handle it. Even though we defaulted, this 
option was overridable.

Now, every signed integer requires explicit specification as to how to 
decode it. UInts and floats don't need this explicit specification due 
to a lack of ambiguity.