Rework the typing structure.
To match the pint/sint/fixed/sfixed pragmas used for marking how a field 
in an object is encoded, PIntWrapped, SIntWrapped, UIntWrapped, 
FixedWrapped, and SFixedWrapped has been added for primitives outside of 
objects.

In order to keep with the requirement of not needing to specify bits, 
templates have been added to convert X to Y, as well as an X value to a 
Y value.

VarIntSubType was moved to Reader. The pragmas are enough for Writer. 
Reader also defines a new Protobuf error and removes all raises of the 
raw ProtobufError.

Also removes IntegerTypes, which was just being aliased to VarIntTypes 
and used as that, and removes RecognizedTypes. RecognizedTypes was just 
used to detect what's implicitly length delimited. The definition of 
LengthDelimitedTypes has been updated to compensate. It should be noted 
object parsing has broken because of that, yet due to the scope of these 
changes and the fact this isn't being used anywhere, I want to get this 
commit committed now.

Finally, legacy types have been explicitly marked as legacy.