Finish the Reader rewrite.
It's not completely finished, as it'd benefit from increased checks. As 
of right now, I believe it's possible to provide either a VarInt or a 
Fixed for an uint field. While this won't decode the wrong value, it 
means two different objects will be treated with structural equality.

Enables nested object reading from a generated code stand point.

Fixes field number detection checking.

Adds a pragma for UInt fields (previously there was only Fixed, a 
variation of UInt).

Fixes 
https://github.com/kayabaNerve/nim-protobuf-serialization/issues/2.

Removes a test case. It encoded X, and then wrote X as a length 
delimited field, then expected the Reader to ignore the outer object for 
X alone.