Port the libp2p VarInt test.
It doesn't test the libp2p modified VarInt variant, and it does 
partially fail tests it should pass. I believe that's due to the test 
being written for a different system though.

This test frequently violates type safety, casting uints to ints which 
are supposed to be negative and more. I understand the test was written 
for a different system, and makes sense in the context of this system, 
but porting it is still a WIP.