More work on Options.
This has truly reached a tipping point for 
https://github.com/kayabaNerve/nim-protobuf-serialization/issues/6. By 
combining that issue with 
https://github.com/kayabaNerve/nim-protobuf-serialization/issues/8, I 
think this library is salavageable. That said, without those, this will 
be impossible to maintain and not submittable.

I want to finish Option support so there's an established baseline. From 
there, I'll do some initial cleanup, and then finally move to the new 
setters and finish cleaning.