Merge PInt and UInt. Add LibP2P Int pragmas/
PInt and UInt follow the same encoding rules for positive values, except 
the sign bit in PInt is used for the value bit in UInt. There's also no 
way this can be ambiguous.

I do worry about the fact it may encourage PInt usage, when SInt should 
be used for signed values (when deciding which type to use, not when 
matching compatibility). That said, it's incorrect to group UInt with 
SInt due to vastly different encoding methods.